+++ app/controllers/dorsale/customer_vault/links_controller.rb
@@ -8,25 +8,25 @@ class Dorsale::CustomerVault::LinksController < ::Dorsale::CustomerVault::Applic
   ]
 
   def new
-    authorize! :update, @person
+    authorize @person, :update?
 
-    @link   ||= model.new
-    @people ||= current_user_scope.people
-  end
-
-  def edit
-    authorize! :update, @person
-
-    @link = model.find(params[:id])
+    @link   ||= scope.new
+    @people ||= person_policy_scope
   end
 
   def create
-    authorize! :update, @person
+    authorize @person, :update?
 
     params = link_params
     bob = params[:bob].split("-")
 
-    @link ||= model.new(title: params[:title], alice_id: @person.id, alice_type: @person.class.to_s, bob_id: bob[1], bob_type: bob[0])
+    @link ||= scope.new(
+      :title      => params[:title],
+      :alice_id   => @person.id,
+      :alice_type => @person.class.to_s,
+      :bob_id     => bob[1],
+      :bob_type   => bob[0],
+    )
 
     if @link.save
       flash[:notice] = t("messages.links.create_ok")
@@ -36,10 +36,16 @@ class Dorsale::CustomerVault::LinksController < ::Dorsale::CustomerVault::Applic
     end
   end
 
-  def update
-    authorize! :update, @person
+  def edit
+    authorize @person, :update?
 
-    @link = model.find(params[:id])
+    @link = scope.find(params[:id])
+  end
+
+  def update
+    authorize @person, :update?
+
+    @link = scope.find(params[:id])
 
     if @link.update(link_params)
       flash[:notice] = t("messages.links.update_ok")
@@ -50,14 +56,14 @@ class Dorsale::CustomerVault::LinksController < ::Dorsale::CustomerVault::Applic
   end
 
   def destroy
-    authorize! :update, @person
+    authorize @person, :update?
 
-    @link = model.find(params[:id])
+    @link = scope.find(params[:id])
 
     if @link.destroy
-      flash[:notice] = t("messages.links.destroy_ok")
+      flash[:notice] = t("messages.links.delete_ok")
     else
-      flash[:alert] = t("messages.links.destroy_error")
+      flash[:alert] = t("messages.links.delete_error")
     end
 
     redirect_to back_url
@@ -72,6 +78,10 @@ class Dorsale::CustomerVault::LinksController < ::Dorsale::CustomerVault::Applic
 
   def back_url
     polymorphic_path(@person) + "#links"
+  end
+
+  def scope
+    policy_scope(model)
   end
 
   def load_linkable
