+++ app/controllers/dorsale/flyboy/tasks_controller.rb
@@ -11,9 +11,9 @@ class Dorsale::Flyboy::TasksController < ::Dorsale::Flyboy::ApplicationControlle
   ]
 
   def index
-    authorize! :list, model
+    authorize model, :list?
 
-    @tasks ||= current_user_scope.tasks
+    @tasks ||= scope.all
 
     @order ||= sortable_column_order do |column, direction|
       case column
@@ -71,24 +71,22 @@ class Dorsale::Flyboy::TasksController < ::Dorsale::Flyboy::ApplicationControlle
   end
 
   def show
-    @task = model.find(params[:id])
-
-    authorize! :read, @task
+    authorize @task, :read?
   end
 
   def new
-    @task = current_user_scope.new_task
+    @task ||= scope.new
     @task.taskable_guid = params[:taskable_guid]
 
     set_owners
 
-    authorize! :create, @task
+    authorize @task, :create?
   end
 
   def create
-    @task ||= current_user_scope.new_task(task_params)
+    @task ||= scope.new(task_params)
 
-    authorize! :create, @task
+    authorize @task, :create?
 
     if @task.save
       flash[:success] = t("messages.tasks.create_ok")
@@ -101,16 +99,15 @@ class Dorsale::Flyboy::TasksController < ::Dorsale::Flyboy::ApplicationControlle
   end
 
   def edit
-    authorize! :update, @task
+    authorize @task, :update?
 
     set_owners
   end
 
   def update
-    authorize! :update, @task
+    authorize @task, :update?
 
-
-    if @task.update_attributes(task_params)
+    if @task.update(task_params)
       flash[:success] = t("messages.tasks.update_ok")
       redirect_to back_url
     else
@@ -120,7 +117,7 @@ class Dorsale::Flyboy::TasksController < ::Dorsale::Flyboy::ApplicationControlle
   end
 
   def destroy
-    authorize! :delete, @task
+    authorize @task, :delete?
 
     if @task.destroy
       flash[:success] = t("messages.tasks.delete_ok")
@@ -132,7 +129,7 @@ class Dorsale::Flyboy::TasksController < ::Dorsale::Flyboy::ApplicationControlle
   end
 
   def complete
-    authorize! :complete, @task
+    authorize @task, :complete?
 
     @task_comment ||= @task.comments.new(
       :progress    => 100,
@@ -151,6 +148,8 @@ class Dorsale::Flyboy::TasksController < ::Dorsale::Flyboy::ApplicationControlle
   end
 
   def snooze
+    authorize @task, :snooze?
+
     @task.snooze
 
     if @task.save
@@ -163,6 +162,8 @@ class Dorsale::Flyboy::TasksController < ::Dorsale::Flyboy::ApplicationControlle
   end
 
   def summary
+    authorize model, :list?
+
     setup_tasks_summary
   end
 
@@ -170,6 +171,10 @@ class Dorsale::Flyboy::TasksController < ::Dorsale::Flyboy::ApplicationControlle
 
   def model
     ::Dorsale::Flyboy::Task
+  end
+
+  def scope
+    policy_scope(model)
   end
 
   def default_back_url
@@ -193,12 +198,12 @@ class Dorsale::Flyboy::TasksController < ::Dorsale::Flyboy::ApplicationControlle
   end
 
   def set_objects
-    @task     = model.find params[:id]
-    @taskable = @task.taskable
+    @task     ||= scope.find(params[:id])
+    @taskable ||= @task.taskable
   end
 
   def set_owners
-    @owners ||= current_user_scope.colleagues(@task.taskable)
+    @owners ||= policy_scope(User).all
   end
 
   def permitted_params
